pyproject.toml
[project]
name = "translate-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.5.0",
]
